Code motion: share TFLOG interceptors in unstable/logging
Previously this code was internal in the pf module, but is now lifted so it can be used in top-level
module as well. The entry-point is consolidated as InitLogging.
